Rename TestSimulator -> Simulator
The name TestSimulator causes pytest to raise some warnings when
running the test suite because it marks anything that starts with
`Test` as being a test to be run. It, fortunately, doesn't run it
as a test because it contains an `__init__`, but raises a warning
as though we had incorrectly defined some tests, where in actuality
we didn't want TestSimulator to be marked as a test in the first
place.